modify project document
