+++ src/app/component/project/1_addProject.tsx
@@ -1,30 +1,19 @@
 import { Dispatch, SetStateAction } from 'react';
-import { useState } from 'react'
+import { Project } from '@/types/types';
 import { auth, db } from '../../../../firebase';
 import { collection, doc, setDoc} from 'firebase/firestore';
 import { useAuth } from '../auth/authContext';
-import Image from "next/image";
+import { useSelector } from 'react-redux';
+import { AppDispatch, RootState } from '@/store';
+import { useDispatch } from 'react-redux';
+import { addProjects } from '@/features/projectsSlice';
 
+const AddProject = () => {
 
-interface Project {
-    id: string;
-    uid: string;
-    projectTitle: string;
-    projectStatus: string;
-    projectMember: string[];
-    projectDateStart: string;
-    projectDateEnd: string;
-    projectOwner: string | undefined;
-    createdAt: string;
-}
+    //Store data
+    const dispatch:AppDispatch = useDispatch();
 
-interface AddProjectProps {
-setProjects: Dispatch<SetStateAction<Project[]>>;
-projects: Project[];
-}
-
-const AddProject: React.FC<AddProjectProps> = ({setProjects,projects}) => {
-    
+    //project data
     const {currentUser,loadingUser} = useAuth();
     const projectTitle = '';
     const projectStatus = 'Unstarted';
@@ -32,38 +21,34 @@ const AddProject: React.FC<AddProjectProps> = ({setProjects,projects}) => {
     const projectDateStart = new Date().toISOString();//應該直接設置今天
     const projectDateEnd = '';
     const projectOwner: string | undefined = currentUser?.email ?? undefined;
+    const projectColor =  '';
 
-    
-    
+
 
     //新增一個project
     const handleAddProject = async() =>{
         if(!loadingUser){
             if(currentUser){
                 const newDocRef = doc(collection(db, 'project'));
-                const newProject = {
-                    uid:
-                        currentUser.uid,
-                        projectTitle,
-                        projectStatus,
-                        projectMember,
-                        projectDateStart,
-                        projectDateEnd,
-                        projectOwner,
-                        createdAt: new Date().toISOString()
+                const newProject:Project = {
+                    id: newDocRef.id,
+                    uid:currentUser.uid,
+                    projectTitle,
+                    projectColor,
+                    projectStatus,
+                    projectMember,
+                    projectDateStart,
+                    projectDateEnd,
+                    projectOwner,
+                    createdAt: new Date().toISOString()
                 }
-                await setDoc(newDocRef, newProject)
-                console.log('已新增一個Project了', newDocRef.id)
-
-                setProjects([ { id: newDocRef.id, ...newProject },...projects]);
                 
+                dispatch(addProjects({newDocRef, newProject}));
             }
-            console.log('no currentUser')
         }
         return;
     }
 
-    
     
     return(
         <> 
@@ -73,7 +58,6 @@ const AddProject: React.FC<AddProjectProps> = ({setProjects,projects}) => {
 
         </>
     )   
-
 
 }
