+++ src/app/component/project/1_projectList.tsx
@@ -1,3 +1,4 @@
+"use client"
 import {useEffect, useState} from 'react';
 import { collection, query, where, getDocs, doc,updateDoc, deleteDoc,getDoc } from 'firebase/firestore';
 import AddCategory from "./2_addCategory";
@@ -5,26 +6,21 @@ import { db } from '../../../../firebase';
 import CategoryContent from './2_categoryContent'
 import MemberList from './member/member_List'
 import Image from 'next/image';
-
-
+import { Category, Project } from '@/types/types';
 import 'flatpickr/dist/themes/material_green.css'
 
-interface Project {
-    id: string;
-    uid: string;
-    projectTitle: string;
-    projectStatus: string;
-    projectMember: string[];
-    projectDateStart: string;
-    projectDateEnd: string;
-    projectOwner: string | undefined;
-    createdAt: string;
-}
+import { useSelector } from 'react-redux';
+import { AppDispatch, RootState } from '@/store';
+import { useDispatch } from 'react-redux';
+import CategoryContentTest from './2_categoryContent';
+import { deleteCategories, fetchCategories, updateCategories } from '@/features/categoriesSlice';
+import AddCategoryTest from './2_addCategory';
+import { deleteProjects, updateProjects } from '@/features/projectsSlice';
+import { updateTasksAsync } from '@/features/tasksSlice';
+import { useAuth } from '../auth/authContext';
 
 interface ProjectListProps {
     project: Project;
-    OnDelete: () => void;
-    OnUpdate: (id: string, updatedData: Partial<Project>) => void;
 }
 
 const getStatusColor = (status: string): string => {
@@ -40,20 +36,34 @@ const getStatusColor = (status: string): string => {
     }
 }
 
-const ProjectList: React.FC<ProjectListProps>  = ({project,OnDelete,OnUpdate}) => {
+const ProjectList: React.FC<ProjectListProps>  = ({project}) => {
+    //----------------store data目前使用者的project---------------------
+    const dispatch:AppDispatch = useDispatch();
+    // task
+    const categories = useSelector((state:RootState) => state.categories.categories);
+    const loading = useSelector((state:RootState) => state.categories.loading);
+ 
+    const { currentUser } = useAuth();
+    //Store data
+    
+    
+
+    //category 資料
+    const [projectCategories, setProjectCategories] = useState<Category[]>([])
+
     //set project資料 
+    const projectId = project.id
     const [title, setTitle] = useState(project.projectTitle);
     const [status, setStatus] = useState(project.projectStatus);
     const [statusColor, setStatusColor] = useState(getStatusColor(project.projectStatus));
     //set members資料
-    const [members, setMembers] = useState(Array.isArray(project.projectMember) ? project.projectMember : [project.projectMember]); 
+    const members= Array.isArray(project.projectMember) ? project.projectMember : [project.projectMember]; 
     const [assignedMembers, setAssignedMembers] = useState<string[]>([]);
     const [notAssignedMembers, setNotAssignedMembers] = useState<string[]>([]);
     //set date資料
     const [startDate, setStartDate] = useState(typeof project.projectDateStart === 'string' ? project.projectDateStart.slice(0, 10) : '');
     const [endDate, setEndDate] = useState(typeof project.projectDateEnd === 'string' ? project.projectDateEnd.slice(0, 10) : '');
-    //category 資料
-    const [categories, setCategories] = useState<any[]>([]); 
+    
    
     //members 資料
     const [showMemberInput, setShowMemberInput] = useState(false)
@@ -63,52 +73,27 @@ const ProjectList: React.FC<ProjectListProps>  = ({project,OnDelete,OnUpdate}) =
     const [isDeleteVisible, setIsDeleteVisible] = useState(false);
     const [isContentVisible,setIsContentVisible] = useState<boolean>(false);
 
-    
-    //fetch category的資料
-    useEffect(()=>{
-        const fetchCategories = async () => {
-            try{
-                const q = query(collection(db,`project/${project.id}/category`));
-                const querySnapshot = await getDocs(q);
-                const fetchedCategories = querySnapshot.docs.map(doc =>({
-                    id:doc.id,
-                    ...doc.data()
-                }));
-                if(fetchedCategories){
-                    setCategories(fetchedCategories);
-                    console.log('fetchedCategories',fetchedCategories)
-                }else{
-                    console.log('這個project沒有categories');
-                }
-            }catch(error){
-                console.log('獲取category出錯',error)
-            }
+   
+    //fetch category
+    useEffect(() => {
+        const filteredCategories = categories.filter(category =>
+            category.projectId === projectId 
+        )
+        setProjectCategories(filteredCategories)
+        
+    },[categories,loading,projectId,dispatch])
 
-        } 
-        fetchCategories()
-    },[project.id])
-        //----------------------- member 處理----------------------------
-    
+    //----------------------- member 處理----------------------------
 
     //新增會員權限
     const handleAddMember = async() => {
         if(newMember && !members.includes(newMember) ){
-            try{
-                const newMemberRef = doc (db, 'project',project.id);
+                const projectDef = doc (db, 'project',projectId);
                 const updatedMembers = [...members, newMember];
-                await updateDoc(newMemberRef, {
-                    projectMember: updatedMembers
-                });
-
-                setMembers(updatedMembers);
+                const updatedData = {projectMember: updatedMembers}
+                dispatch(updateProjects({projectDef,updatedData,projectId}))
                 setNewMember('');
                 setShowMemberInput(false);
-            } catch(error){
-                console.log('新增成員時出錯',error);
-            }
-           
-            
-
         }
     }
 
@@ -116,12 +101,10 @@ const ProjectList: React.FC<ProjectListProps>  = ({project,OnDelete,OnUpdate}) =
     const handleDeleteMember = async(memberToRemove:string) => {
         try{
             const updatedMembers = members.filter(member => member !== memberToRemove)
-            const newMemberRef = doc (db, 'project',project.id);
-            await updateDoc(newMemberRef, {
-                projectMember: updatedMembers
-            });
-            setMembers(updatedMembers);//更新local端
-
+            const projectDef = doc (db, 'project',projectId);
+            const updatedData = {projectMember: updatedMembers}
+            dispatch(updateProjects({projectDef,updatedData,projectId}))
+            
             //獲取這個project的所有task
             const categorySnapshot = await getDocs(collection(db, `project/${project.id}/category`));
             let tasks:any[] = [];
@@ -139,12 +122,14 @@ const ProjectList: React.FC<ProjectListProps>  = ({project,OnDelete,OnUpdate}) =
                 const taskRef = doc(db,`project/${project.id}/category/${task.categoryId}/task/${task.taskId}`)
                 const taskDoc = await getDoc(taskRef);
                 const taskData = taskDoc.data();
+                const taskId = task.id
                 if(taskData && taskData.taskAssign.includes(memberToRemove)){
+
                     const updatedTaskAssign = taskData.taskAssign.filter((member:string)=> member !== memberToRemove);
-                    await updateDoc(taskRef,{
-                        taskAssign: updatedTaskAssign
-                    });
-                    console.log('成功刪除taskassign成員', updatedTaskAssign)
+                    const taskRefString = `project/${project.id}/category/${task.categoryId}/task`;
+                    const updatedData = {taskAssign: updatedTaskAssign};
+
+                    dispatch(updateTasksAsync({taskRefString,updatedData,taskId}))
                     setAssignedMembers(updatedTaskAssign);
                     setNotAssignedMembers(updatedMembers.filter(member => !updatedTaskAssign.includes(member))) 
                     
@@ -161,63 +146,33 @@ const ProjectList: React.FC<ProjectListProps>  = ({project,OnDelete,OnUpdate}) =
         setAssignedMembers(newAssignedMembers);
         setNotAssignedMembers(newNotAssignedMembers);
     }
-    //-----Category-------
-    //刪除Category
-    const onDeleteCategory = async(categoryId:string) => {
-        try{
-            const categoryDef = doc(db, `project/${project.id}/category/${categoryId}`);
-            await deleteDoc(categoryDef);
-            setCategories(prevCategories => prevCategories.filter(category => category.id !== categoryId))
-        }catch(error){
-            console.log('刪除category時錯誤',error)
-        }
-    }
-
-    //即時儲存更新category
-    const handleUpdateCategory = async(categoryId:string, updatedData:any) =>{
-        try{
-            const categoryDef = doc(db, `project/${project.id}/category/${categoryId}`);
-            await updateDoc(categoryDef, updatedData);
-            setCategories(prevCategories => prevCategories.map(category =>
-                category.id === categoryId ? { ...category, ...updatedData } : category
-            ));
-        }catch(error){
-            console.log('更新category時錯誤',error);
-        }
-
-    }
-
+    
 
     //--------------------------編輯project內容---------------------------
-    //編輯標題
-    const handleTitleBlur = () =>  {
-        OnUpdate(project.id, {projectTitle: title})
-        console.log('確認project ID',project.id)
-    }
-    //編輯狀態
-    const handleStatus = () => {
-        OnUpdate(project.id, {projectStatus: status})
-    }
+     //刪除project
+     const onDeleteProject = async() => {
+        const projectDef = doc(db, 'project', projectId);
+        dispatch(deleteProjects({projectDef, projectId}))
+       
+    };
+    //更新project的內容
+    const handleUpdateProject = async(updatedData:Partial<Project>) => {
+        const projectDef = doc(db, 'project', projectId);
+        dispatch(updateProjects({projectDef, updatedData, projectId}))
+    };
+    
     //轉換狀態顏色
     const handleStatusColor = (e:React.ChangeEvent<HTMLSelectElement>) => {
         const value = e.target.value;
         setStatus(value)
         setStatusColor(getStatusColor(value))
     }
-   
-    //編輯開始日程
-    const handleStartDate = () =>{
-        OnUpdate(project.id, {projectDateStart: startDate})
-    }
-    //編輯結束日期
-    const handleEndDate = () => {
-        OnUpdate(project.id, {projectDateEnd: endDate})
-    }
     //-------------------------切換狀態-------------------------------
 
     const toggleContent = () => {
         setIsContentVisible(!isContentVisible);
     }
+
     return (
         <>
             <div className='project'>
@@ -233,7 +188,13 @@ const ProjectList: React.FC<ProjectListProps>  = ({project,OnDelete,OnUpdate}) =
                                 className={`project-toggleImg ${isContentVisible? 'rotate':''}`} 
                                 src="/images/toggle.svg" alt="project toggle" width={20} height={20}/>
                         </button>
-                        <input className='project-title' placeholder='Title' value={title} onChange={(e)=>setTitle(e.target.value)} onBlur={handleTitleBlur}></input>
+                        <input 
+                            className='project-title' 
+                            placeholder='Title' 
+                            value={title} 
+                            onChange={(e)=>setTitle(e.target.value)} 
+                            onBlur = {(e) => handleUpdateProject({projectTitle: e.target.value})}
+                        ></input>
                         
                         
                         
@@ -241,7 +202,11 @@ const ProjectList: React.FC<ProjectListProps>  = ({project,OnDelete,OnUpdate}) =
 
                         <div className='project-status'>
                             <label className="project-status-color" htmlFor="status"  style={{backgroundColor:statusColor}}></label>
-                            <select id="status" value={status} onChange={handleStatusColor} onBlur={handleStatus}>
+                            <select id="status" 
+                                value={status} 
+                                onChange={handleStatusColor} 
+                                onBlur = {(e) => handleUpdateProject({projectStatus: e.target.value})}
+                                >
                                 <option value="Unstarted">Unstarted</option>
                                 <option value="Processing">Processing</option>
                                 <option value="Done">Done</option>
@@ -294,21 +259,21 @@ const ProjectList: React.FC<ProjectListProps>  = ({project,OnDelete,OnUpdate}) =
                                 type = "date" 
                                 value = {startDate}
                                 onChange = {(e) => setStartDate(e.target.value)}
-                                onBlur = {handleStartDate}
+                                onBlur = {(e) => handleUpdateProject({projectDateStart: e.target.value})}
                             />
                             ~
                             <input 
                                 type="date" 
                                 value={endDate}
                                 onChange={(e) => setEndDate(e.target.value)}
-                                onBlur = {handleEndDate}    
+                                onBlur = {(e) => handleUpdateProject({projectDateEnd: e.target.value})}
                             />
                         </div>
                         
                         <div className='project-delete' onClick={() => setIsDeleteVisible(prev => !prev)}>⋮ 
                             {isDeleteVisible?
                             <>
-                                <div className='project-delete-block' onClick={OnDelete}>
+                                <div className='project-delete-block' onClick={onDeleteProject}>
                                 <Image  src="/images/delete.svg" alt="project delete" width={20} height={20}/>
                                 </div>
                                 <div className="overlay"></div>
@@ -330,19 +295,15 @@ const ProjectList: React.FC<ProjectListProps>  = ({project,OnDelete,OnUpdate}) =
                         <div className='project-content-addCat'>
                             <AddCategory
                                 projectId={project.id}
-                                setCategories={setCategories}
-                                categories={categories}
                             />
                         </div>
                         <div className='project-content-taskblock'>
-                            {categories?.map(category => (
-                                        <CategoryContent 
-                                        key={category.id}
-                                        category={category}
-                                        OnDelete={()=>onDeleteCategory(category.id)}
-                                        OnUpdate={handleUpdateCategory}
-                                        members={members}
-                                        updatedMembers={handleMembersChange}
+                            {projectCategories.map(category => (
+                                    <CategoryContent
+                                    key={category.id}
+                                    category={category}
+                                    members={members}
+                                    updatedMembers={handleMembersChange}
                                     />
                                     ))}
                         </div>
