+++ src/app/component/project/2_addCategory.tsx
@@ -1,61 +1,45 @@
-import { Dispatch, SetStateAction } from 'react';
+
 import { useAuth } from '../auth/authContext';
 import { auth, db } from '../../../../firebase';
 import { collection, doc, setDoc} from 'firebase/firestore';
-import { useState } from 'react'
-import Image from 'next/image';
+import { Category } from '@/types/types';
 
-interface Category {
-    id: string;
-    uid: string;
-    categoryTitle: string;
-    createAt: string;
-    projectId: string;
-}
-
+import { AppDispatch} from '@/store';
+import { useDispatch } from 'react-redux';
+import { addCategories } from '@/features/categoriesSlice';
 
 interface AddCategoryProps {
     projectId: string;
-    setCategories: Dispatch<SetStateAction<Category[]>>;
-    categories: Category[];
 }
 
 
-const AddCategory:React.FC<AddCategoryProps>= ({projectId, setCategories, categories}) => {
+const AddCategory:React.FC<AddCategoryProps>= ({projectId}) => {
+    
     const {currentUser,loadingUser} = useAuth();
+    //Store data
+    const dispatch:AppDispatch = useDispatch();
+
     const categoryTitle  = '';
 
     const handleAddCategory = async() => {
         if(!loadingUser && currentUser){
-            try{
-                const newDocRef = doc(collection(db, `project/${projectId}/category`));
-                const newCategory = {
-                    uid:
-                        currentUser?.uid,
-                        categoryTitle,
-                        createAt: new Date().toISOString(),
-                        projectId: projectId
-                }
-                await setDoc(newDocRef, newCategory);
-
-                setCategories(categories => (categories? [{id: newDocRef.id,...newCategory},...categories]:[{id: newDocRef.id,...newCategory}]));
-                
-            }catch(error){
-                console.log('fetch category的時候錯誤',error);
+            const newDocRefCat = doc(collection(db, `project/${projectId}/category`));
+            const newCategory:Category = {
+                id: newDocRefCat.id,
+                uid:currentUser?.uid,
+                categoryTitle,
+                createAt: new Date().toISOString(),
+                projectId: projectId
             }
+            dispatch(addCategories({newDocRefCat,newCategory}))
         }
-
     }
-    
     return (
-
-    <> 
-
-    <div className='category-button' onClick={handleAddCategory}>
-        +<div className='category-text'>Category</div>
-    </div>
-    
-    </>
+        <> 
+            <div className='category-button' onClick={handleAddCategory}>
+                +<div className='category-text'>Category</div>
+            </div>
+        </>
     )
