+++ src/app/pages/project/page.tsx
@@ -1,14 +1,13 @@
 "use client"
-import ProjectMainTest from "@/app/component/project/projectTest/0_projectMainTest";
-import "./project.scss"
 import ProjectMain from "@/app/component/project/0_projectMain";
+import "./project.scss"
 
 
 const ProjectHome = () => {
     
   return (
     <>
-     <ProjectMainTest/>
+     <ProjectMain/>
     </>
